Silence rm output

+++ nvm.sh
@@ -2695,11 +2695,11 @@ nvm() {
           if [ -f $installed_flag ]; then
             npm install --color false -g grunt-cli bower
 
-            rm $installed_flag 2>&1 1>/dev/null
+            [ -f $installed_flag ] && rm $installed_flag 2>&1 1>/dev/null
             return
           fi
 
-          rm $installed_flag 2>&1 1>/dev/null
+          [ -f $installed_flag ] && rm $installed_flag 2>&1 1>/dev/null
           tries=`expr $tries + 1`
         done
